requirements.txt
# Backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0

# Code analysis & correction
autopep8>=2.0.0
flake8>=6.0.0
